+++ src/modules/CommunityProfile/index.ts
@@ -1,3 +1,3 @@
 export * from './CommunityBgImage.client';
-export * from './CommunityLogoImage.client';
-export * from './CommunityProfile.client';
+export * from './CommunityLogoImage';
+export * from './CommunityProfile';
